refactor: 순위, 당첨금, 수익률 관련 메서드 LottoStatistics 로 이동
